apps/api/pyproject.toml
[project]
name = "photoprune-api"
version = "0.0.0"
description = "PhotoPrune FastAPI service skeleton"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.0",
    "uvicorn>=0.30.6",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "pillow>=10.3.0",
]

[dependency-groups]
dev = [
    "ruff==0.14.11",
    "black==24.10.0",
    "mypy==1.19.1",
    "pytest==8.4.2",
    "pytest-cov==5.0.0",
    "httpx==0.28.1",
    "pip-audit==2.10.0",
]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "B", "W", "UP"]
lint.ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_unused_ignores = true
show_error_codes = true
mypy_path = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=xml --cov-report=term"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
